EviLTusuf_V2.exe
Destroys the computer with regedit keys. It installs a lot of files on the computer. It infects your disk. Windows PE also blocks things like administrator account, safe mode, cmd, powershell, taskmgr, sethc, etc. It deletes a lot of files and makes the computer unrepairable. PLEASE OPEN THIS SOFTWARE ON A VIRTUAL MACHINE, NOT ON YOUR ERÇEK COMPUTER!